+++ CHANGELOG.md
@@ -3,6 +3,7 @@ LUYA CHANGELOG
 
 1.0.0-beta1 (in progress)
 -------------------------
+- **[BC BREAK]** `#344` zaa-checkbox-array value describer `id` is replaced by `value` to be consistent
 - added ability to register asset class inside a block
 - added ability to create block ajax requests with callbacks.
 - added ability to add rewrite if not existing in same navigation level.